Fixed various issues
I'll clean this all up later when I'm fresh.

There were several issues with the code before, mostly surrounding
exactly when items were stored on disk and fetchable by readers.

I had to pull the checksumming code in on the writer-side of things
because it can't send a memory block onto the memory clearer until it is
certain the buffer areas involved are truly written to disk.

I want to make the code clearer and get better separation on areas of
interest, but I also don't want to slow things down too much. So, I'll
wait until I have a clear head.