Switched to micro second timestamps and...
Switched to micro second timestamps and working on push replication.

The timestamps for the user of this package are now in microseconds to
make room for a set of internally used util bits for deletion markers
and "do not replicate" markers (needed for push replication) and maybe
more bits later. With this change, the package user now gets the
timestamp he gives in all cases instead of how it was before where
deletes "oddified" timestamps and non-deletes "evenified" them. All the
bit mangling is now hidden from the outside package user.

This allows a new "do not replicate" bit that indicates an item that was
stored in a valuestore even though the valuestore wasn't responsible for
it (according to the ring) and then was succesfully push replicated to
another valuestore. This "do no replicate" action is stored to disk like
any other so that restarting the service won't cause re-replication of
everything it has ever seen even if it is no longer responsible for some
of the data and had succesfully push replicated it "off".

However, there is still refactoring of the recovery routines to account
for this new reality.

Currently, an item could be received and stored, then deemed not locally
responsible and push replicated off and the "do not replicate" marker
stored, and then later the item come back in again after a ring change
that makes the valuestore responsible at that time, and therefore stored
again. In this case, the item is stored in the sequence "actual value /
do not replicate / actual value" which should result in the actual value
being active at the end of recovery.

However, the recover routines load up data from disk as fast as possible
and have no guarantee on the order the item actions will be set in the
valuelocmap (which was fine before). It currently loads a block of
actions and sends that block to a channel, a channel that many workers
read from concurrently.

Instead, I need to refactor recovery to load actions and send them to
various channels based on the key with one worker reading from one
channel, thus guaranteeing that loaded actions for any given item will
get applied in the order they were written to disk.