Reworked Opt to Config
After discussing with @pandemicsyn at length, we decided this was a
better approach. Using a public struct makes it just slightly harder for
the easy cases, but a lot easier for the complicated cases.

For example, the previous public global Opt funcs are completely
undiscoverable using reflection, but the fields inside the new public
Config struct can be used with reflect.